Adding some simple postgres commands

+++ src/test/java/jmdbtutorial/postgres/shard/Basic_Postgres_Connection.java
@@ -8,7 +8,6 @@ import java.util.List;
 import java.util.Properties;
 
 import static java.lang.String.format;
-import static junit.framework.TestCase.assertTrue;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -20,18 +19,20 @@ public class Basic_Postgres_Connection {
 
         String url = "jdbc:postgresql://localhost/";
         Properties props = new Properties();
-        props.setProperty("user","postgres");
-        props.setProperty("password","");
+        props.setProperty("user", "postgres");
+        props.setProperty("password", "");
 
 
         try (Connection conn = DriverManager.getConnection(url, props)) {
-            executeSql(conn, "DROP TABLE CUSTOMER;");
-            executeSql(conn, "CREATE TABLE CUSTOMER (id CHAR(255), name CHAR(255));");
+            if (tableExists(conn, "CUSTOMER")) {
+                executeSql(conn, "DROP TABLE CUSTOMER;");
+            }
+            executeSql(conn, "CREATE TABLE CUSTOMER (id VARCHAR, name VARCHAR);");
 
             executeSql(conn, "INSERT INTO CUSTOMER VALUES ('000000001', 'foo bar');");
 
 
-            ResultSet rs =  executeQuery(conn, "SELECT * FROM CUSTOMER ;");
+            ResultSet rs = executeQuery(conn, "SELECT * FROM CUSTOMER ;");
 
             List<String> rows = readRows(rs);
             assertThat(rows.size(), is(1));
@@ -39,7 +40,13 @@ public class Basic_Postgres_Connection {
         }
 
 
+    }
 
+    private static boolean tableExists(Connection conn, String tableName
+    ) throws SQLException {
+        ResultSet rs = conn.prepareStatement(format("SELECT to_regclass('%s');", tableName)).executeQuery();
+        rs.next();
+        return rs.getString(1) != null;
     }
 
     private void printRows(List<String> rows) {
@@ -55,12 +62,12 @@ public class Basic_Postgres_Connection {
         while (rs.next()) {
             String id = rs.getString("id");
             String name = rs.getString("name");
-            rows.add(format("%s \t %s", id.trim(), name.trim()));
+            rows.add(format("%s \t %s", id, name));
         }
         return rows;
     }
 
-    private static  void executeSql(Connection conn, String sql) throws SQLException {
+    private static void executeSql(Connection conn, String sql) throws SQLException {
         PreparedStatement st = conn.prepareStatement(sql);
         st.execute();
     }
